cognex_sap315_platform/cognex_integration/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
opencv-python==4.9.0.80
Pillow==10.0.1
asyncio==3.4.3
websockets==12.0
structlog==24.1.0
pydantic==2.5.3
python-dateutil==2.8.2
pytest==7.4.4